Update the deployment repo and also detect failure
A failed upload was not being detected and causing the job to fail. This
fixes it so that if the maven upload fails we'll fail the build.

Also, at Ed's request there is now a .dev repo for the regular merge
time deb files so we need to push there and not to .release

Change-Id: I08e01979c35c248f6e0fcc6c3b166f2148a9b9f0
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>