Change honeycomb-settings UUID
JJB doesn't handle the latest config file provider IDs for objects
properly as it expects them all to be prefixed with
org.jenkinsci.plugins.configfiles.maven.MavenSettingsConfig otherwise it
believes the information to be a file path (which doesn't work!) As
such, the honeycomb-settings ID has been fixed up to be more inline with
what JJB expects / requires (until we can get it patched). This required
a restart of Jenkins to take effect.

Change-Id: I60092b1a229d498d672b60fef9554fd7a16d5b60
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>