Move JJB jobs to dedicated static slave
Zuul has a problem where it can't control jobs that are tied to the
master. As we need JJB jobs to also be able to actually do some work on
the the Zuul system we're statically tying it into Jenkins with a
requirement that only jobs specifically targetting it can run there.

Change-Id: I5a689056edbfa2e154b88d6b3df0a9b97fa17192
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>