Update patch set 2
Patch Set 2: Code-Review+1

So we are currently doing this is the Jenkins spin-up script (or something functionally the same). If you would like I'll merge it in, but I don't see a reason to rebuild the image to get it added at this time.

Patch-set: 2
Reviewer: Gerrit User 1 <1@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>
Label: Code-Review=+1