Add vpp jenkins job to build vpp and run csit tests.
To start integrating with VPP verify job, start csit tests for each VPP
verify run. Until artifacts passing is solved between jenkins upstream
and downstream jobs, build the VPP project again in this job. Then
check out csit project, copy the deb packages to the cloned dir and
start the testcase.

Prevent the job non-voting, so that it does not polute VPP reviews with
erroneous results. We'll switch to normal voting after all potential
problems are ironed out, and the job is stable.

Change-Id: I9e35e19a1d8b2a64b8e65473fef93e0b537eded2
Signed-off-by: Stefan Kobza <skobza@cisco.com>