Make subnodes easily identifiable
When nodepool brings online a node and subnodes they are presently all
named the same way. This makes it difficult to pinpoint the leader node
of a set if we need to do any work directly against OpenStack. This
change will cause sub-nodes to have the prefix of np-sub- and the rest
of the name from the leader node making it clear.

Change-Id: I536b2e4437a93f118436e8eeb6869a2f691be0f6
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>