Update patch set 1
Patch Set 1: Code-Review-1

This is not the best way to go about getting a centos image in play. The current basebuild image should be changed to be capable of working on either ubuntu or centos.

This means taking all the installation scripting and moving it to an external script first of all. Having it all in a single Vagrant file is nice for distributing to folks, but it actually makes working on the Vagrant harder.

Examples of multi-distro aware configuration can be seen in the OpenDaylight releng vagrants.

Patch-set: 1
Reviewer: Gerrit User 1 <1@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>
Label: Code-Review=-1