Update patch set 1
Patch Set 1:

> 
 > 1) Moving from zuul back to Gerrit Trigger is not recommended.
 > We're working on deprecated Gerrit Trigger as it is becoming an
 > untenable piece of management overhead since job updates do not
 > refresh / lose their triggers without manual intervention
 > 

Andy, Zuul has proven itself quite unstable for us during the time we've been using it. I understand the need to move to one system rather than 2 (zuul AND gerrit trigger), therefore I'd like to propose following: let's have this job moved to Jenkins for near future (~month), and when Zuul is believed to be stable, we can switch all three jobs (verify, merge and this one) to Zuul system.


VPP jobs are still written with Jenkins triggers, so we are not bringing "old style" job to "new style" job set anyway. All of this comes from promise we gave to VPP community that we are going to have some CSIT testing in two weeks. In fact it is doable today, but we see many problems with Zuul. Going forward with unstable approach for first attempt would shed bad light on the project as well as infrastructure.

Patch-set: 1