Update patch set 2
Patch Set 2:

> > The export TEST_TAG part should go inside of the
 > includeraw-vpp-verify-performance.sh
 > > script.
 > 
 > I am confused now. C.J and Andrew is saying it should be in yaml.
 > Can you please confirm? Thank you.

The reason it needs to go in the ...performance.sh file is because we need to uppercase the TYPE parameter since yaml sends it in lowercase. I forgot to mention that we also need to create a type-parameter macro before it can be used. Add a section like this to your YAML.

- parameter:
    name: type-parameter
    parameters:
        - string:
            name: TYPE
            default: '{type}'
            description: "TYPE parameter if not given by trigger"

This creates an environment variable TYPE which you can use for the short|long part of your parameter. In the ....performance.sh script we need to uppercase it so we do some bash magic with:

 export TEST_TAG="PERFTEST_${TYPE^^}"

It should make TEST_TAG be for example PERFTEST_LONG.

Hope this helps.

Patch-set: 2
Reviewer: Gerrit User 6 <6@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>
Label: Code-Review=0