Query RPM files directly using rpm for upload info
Instead of trying to figure out all the maven fields based upon the
filename with possible failures that this can cause, instead, use the
rpm command itself to query the built artifacts for the needed
information.

Also, .src.rpm should be the canonical name for source RPM files. Make
sure that .srpm and .src.rpm are handled properly and they normalize to
.src.rpm in nexus.

Change-Id: I6b35c15e8e6b189a2144fe3cfb991e467933ad46
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>