Update patch set 4
Patch Set 4:

what is the purpose of this script?  To ensure that the same code is being executed as we specify in the vpp-build.sh script?

If this is the purpose, we should also grab the sum of /vagrant/lib/bootstrap-functions.sh, since most of the work is done from here.

Patch-set: 4
Label: Code-Review=0