Use correct server-name Primary
Perhaps 'Default' was the name at one point and then got renamed to
'Primary' and we forgot to update the configuration. Default is not
currently a valid server-name so resulted in Gerrit Trigger not watching
any valid Gerrit instance for the event-stream.

The current name of the Gerrit server-name is 'Primary' as defined in
Jenkins > Manage Jenkins > Gerrit Trigger.

Change-Id: Ic9481b53424d6fc8358131d384157779a7753fb0
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>