DO NOT MERGE - Setup honeycomb to build and push rpms on merge
This patch depends on the merging of https://gerrit.fd.io/r/#/c/1574/
which must merge before it.

Please also note, while there is a first step to getting the package
pushing scripts into a common sharable place, the vpp script
for pushing has *intentionally* not been changed to use
it yet because of the release later this week as a matter
of risk management.

Change-Id: Ifad185d35f582a62bb750ece3c022a24d2153b4e
Signed-off-by: Ed Warnicke <eaw@cisco.com>