Move from rpm push in merge to rpm push in integration
It turns out that doing rpm push in both merge and integration
job leads to collisions because they have two separate
sequences of BUILD_NUMBERs and those BUILD_NUMBERS are used
for the rpm release number.

So... this patch makes the merge job *stop* pushing rpms,
and should cause the integration job to trigger on both
vpp *and* honeycomb merge job success.

Change-Id: I12fe582ea8a1444096b86c214b3cd3c412bc7591
Signed-off-by: Ed Warnicke <eaw@cisco.com>