Update patch set 3
Patch Set 3: Code-Review+1

The whole point of a code review system is that you can get back old code if you need to ;)

I think the general idea is fine but one thing I'd wonder is does that project ever have a need for a different version of Maven?

If so this patch will make it so that they can no longer pick the version of Maven they want to use. If not then I guess no issue here as we can always force them to use the latest and greatest.

Patch-set: 3
Reviewer: Gerrit User 6 <6@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>
Label: Code-Review=+1