Force Jenkins to not change it's votes for merge
Jenkins by default resets the votes it has in Gerrit to 0. Gerrit 2.13
has problems with this and causes us to catch an error. As such, we
should make sure that Jenkins does not try to change its votes when it
starts during a merge job. We assume that any merge run will be with our
normal mechanically set votes.

Change-Id: Ia45eb7ca605ee904f211065a2877fc0cc1ee5f42
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>