Update sonar to use sonarqube recommended invocation
This patch fixes the Honeycomb merge jobs which have been failing on
Sonar deployment.

Since the post build step for Sonar is deprecated anyway there's no
point continuing to use it so transition to using the sonar method
recommended by SonarQube [0].

The "Prepare SonarQube Scanner environment" is currently not
configurable via Jenkins so explicitly set the Maven Goals directly to
work around this. We should switch back in the future once JJB supports
the setting.

[0]
http://docs.sonarqube.org/display/SCAN/Analyzing+with+SonarQube+Scanner+for+Jenkins

RT: 30789
Change-Id: I5eed53dfb183c248deb277dd583ac1a141fb8151
Also-by: Thanh Ha <thanh.ha@linuxfoundation.org>
Signed-off-by: C.J. Collier <cjcollier@linuxfoundation.org>