Update patch set 4
Patch Set 4:

(1 comment)

Vanessa, are you creating a VM image to run Packer?  I thought we ran Packer on our workstations.  If we are creating a VM to run Packer, I think maybe the VM which runs Packer should maybe be in a different patch, so that we keep things a little simpler.

Patch-set: 4