Update patch set 7
Patch Set 7:

> All images built with these packer templates are for jenkins. As such the name 'jenkins-executor' is actually even more generic and states less to what it does than 'basebuild' does.

It seems reasonable that these templates are all jenkins related, since the files reside in the ci-management repo.  I'm not convinced, though that 'jenkins-executor' is less appropriate.  The argument that 'basebuild' is easily and often confused with 'baseline' was not addressed.  Perhaps we could come to a consensus that involves not starting with 'base'?  Perhaps 'common', 'shared' or 'mutual'?

My understanding is that files under /packer are intended to create VM systems used by projects related to ci-management.  The generic 'jenkins-executor' would be the one used if no more specific image exists.

For instance, the deb_dpdk dependencies are rather vast and share little with the VPP build/test system.  I would think that in the future, this infrastructure could be used to factor the project-specific portions of the 'jenkins-executor' template in to project-specific templates, for instance 'jenkins-executor-vpp' and 'jenkins-executor-deb_dpdk'.

Patch-set: 7