Update patch set 7
Patch Set 7:

Although it is not a blocking concern, I am a little worried that our collab-it packer setup does not support parallel, atomic builds of all target platforms from a single run of packer.

The way we are currently using the same variable name (base_image) to represent the ubuntu-14.04 backing-store, ubuntu-16.04 backing-store and centos backing-store is one of the reasons we cannot use one template to concurrently build all target platforms in one run.

I have been looking through the packer documentation to see if we could set up some kind of associative array of variables such as cloud_user_data['ubuntu1404'] for instance, but it doesn't look like packer is flexible enough to understand such things.

I got around this problem in my build environment by generating a template.  I brought the possibility of using a template generator up with Andy, and he pointed out quite correctly that simplicity is key for these files.   We should not depend on a template generator because it would be one more hurdle to mount for users of our packer system.

Which leaves us in the position of having to execute packer N times where N is the number of build platforms we support.  It would help me sleep better at night if we could find a way to do builds atomically in one run of packer.

Patch-set: 7