Update patch set 7
Patch Set 7:

> > All images built with these packer templates are for jenkins. As
 > such the name 'jenkins-executor' is actually even more generic and
 > states less to what it does than 'basebuild' does.
 > 
 > It seems reasonable that these templates are all jenkins related,
 > since the files reside in the ci-management repo.  I'm not
 > convinced, though that 'jenkins-executor' is less appropriate.  The
 > argument that 'basebuild' is easily and often confused with
 > 'baseline' was not addressed.  Perhaps we could come to a consensus
 > that involves not starting with 'base'?  Perhaps 'common', 'shared'
 > or 'mutual'?
 > 
 > My understanding is that files under /packer are intended to create
 > VM systems used by projects related to ci-management.  The generic
 > 'jenkins-executor' would be the one used if no more specific image
 > exists.
 > 
 > For instance, the deb_dpdk dependencies are rather vast and share
 > little with the VPP build/test system.  I would think that in the
 > future, this infrastructure could be used to factor the
 > project-specific portions of the 'jenkins-executor' template in to
 > project-specific templates, for instance 'jenkins-executor-vpp' and
 > 'jenkins-executor-deb_dpdk'.

I have no problem with names that simple and descriptive. I have an issue with 'jenkins-executor' and particularly having it as part of _all_ the names. Go look at ODL, we have 9 different templates. None of them are called basebuild because the ODL community settled on a name that they wanted to use for it (java-builder) the baseline that's in ODL is just the baseline.sh + reseal.sh scripts that _all_ packer instances are supposed to use.

Having a minion definition per each project in FD.io is fine too for all I care. The reason they don't is because the community just kept layering things onto the one default one I gave them originally and I never saw a reason to get them to break things apart which should be done. However, that's something for later, not now.

Patch-set: 7