Use make verify only after its working and not for CSIT builds.
This patch does the following:

1)  Introduces a new PARAMATER called is-csit-vpp-job
2)  Sets that PARAMETER to 'True' for csit-vpp-jobs
3)  Changes the include-raw-vpp-build.sh so that:
    a)  It uses make verify only if this is not a csit job *and*
        if the patch has as a parent a point from which make verify
        is known to work
    b)  It uses the old build.sh if its building CSIT jobs or the patch
        is from before make verify.

The result is, patches should verify or not on their own merit, not
on the happenstance of coming before or after make verify, and
we don't do the extra make verify work for jobs were CSIT just needs
packages built.

Note: This patch is a fixed version of https://gerrit.fd.io/r/#/c/4083/
and has been tested using the 'DRYRUN' env variable.

Change-Id: Ie6f1e6ce2809e21ea025bccfaec703cd7519e20e
Signed-off-by: Ed Warnicke <hagbard@gmail.com>