Update patch set 1
Patch Set 1: Code-Review-1

Running 'apg-get update' is not the same as 'yum upgrade' at all. It's more equivalent of 'yum check-update'. At most all it is doing is updating the package indexes from the mirrors. Not doing this means that it's very possible for packages that are being installed during jobs via 'apt-get install' will either attempt to pull old versions, or not the ones that they are expecting to install. Particularly in the case of the fd.io deb packages that they are pushing to nexus and installing in other jobs.

Given how the start scripts are currently put together we should not be holding any apt locks at all by the time that Jenkins logs in. The problem with apt locks was entirely due to the auto-update system that Ubuntu has enabled by default and we now forcibly disable. We would remove it but it's stupidly a dependency of cloud-init on Ubuntu 14.04.

Patch-set: 1
Reviewer: Gerrit User 1 <1@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>
Label: Code-Review=-1