Update patch set 3
Patch Set 3:

Maciek,

(re: your comment on Patch set 2),

The code committed in reference to the email that you quoted from Damjan is the implementation of "make dpdk-install-dev", which is what the modification that Peter added uses to ensure that the version specified in vpp/dpdk/Makefile is installed.  I have verified that this works correctly in all cases where vpp/dpdk/Makefile contains an older version, the same version, or a newer version of DPDK.

This will resolve the breakage that occurred when Sergio added the crypto dpdk version.

As far as optimizing the local building of DPDK if there is already the same version available in Nexus, I would say that this logic needs to be added to vpp/dpdk/Makefile, not here in ci-management.

Patch-set: 3