Fix VPP merge job & docs
Change https://gerrit.fd.io/r/#/c/6729 to remove Ubuntu 14.04 builds
from the VPP CI workflow unfortunately also removed the triggers that
cause the some post-merge jobs to run; specifically this meant that
the job that builds VPP documentation was no longer invoked.

The trigger is defined in terms of one earlier post-merge job; in an
ideal world you'd want "any of" or "all of" the upstream jobs to be
listed, but apparently this does not work. The upshot is that the
trigger was defined to run after the Ubuntu 14.04 jobs and that
trigger was removed since it referenced the defunct Ubuntu 14.04.

This should restore that trigger by basing it off of Ubuntu 16.04
instead.

Change-Id: I05dc8738f662ec1c37ee005486b57a7644342333
Signed-off-by: Chris Luke <chrisy@flirble.org>