openSUSE support and some rework to identify OS
Rework to identify OS:
Same approach followed by Makefiles and building scripts in the VPP
repository. It is much cleaner and easier to simply use the environment
variables in /etc/os-release

openSUSE support:
Some of the code added to the scripts can be removed once a separate
patch from Vanessa is merged. The purpose here is to save the changes
made since they will be lost otherwise.

Change-Id: I06a041bbc5f5b7e3fa3510db43acf7d1462ec24c
Signed-off-by: Marco Varlese <marco.varlese@suse.com>