Update patch set 1
Patch Set 1: Code-Review+1

this will continue to throw errors on openstack builder for opensuse but ive true'd it all out so that no builds wont fail.  And we will also continue to get the package list on the container.  If marco comes up with a zypper command that doesnt repo pull ill probably replace yum with that

Patch-set: 1