Flags for parallel test run and arm verify update
Added a script where we can look at the label (flavor) of the jenkins
job and set the number of cores used in build/test for that label.

When setting the number of cores, the label along with the corresponding
hosts must ensure (by properly configuring them in nomad) that the
resources used by jobs on the host won't exceed host's resources.

Change-Id: I138c8ad797df84ebf9a926074a1a7cb5037bd681
Signed-off-by: juraj.linkes <juraj.linkes@pantheon.tech>