Update patch set 4
Patch Set 4: Code-Review-1

The new patch set tries to accommodate these requirements:

Enable unit testing where possible. The regular test run is very long, so we're enabling unit testing only in branches with parallel test execution - master and stable/1810

We want to move to Ubuntu1804. We don't want to have both Ubuntu1804 and Ubuntu1604 running in parallel since we might not have enough resources for that, but maybe we should try to test this - thoughts? In any case, the older releases were not tested with Ubuntu1804, so that's another reason for splitting the jobs into Bionic and Xenial.

I'm sure about the naming of the jobs - please comment.

And the last thing - the master jobs need https://gerrit.fd.io/r/#/c/16335/ and https://gerrit.fd.io/r/#/c/16282/ to pass.

Patch-set: 4
Label: Code-Review=-1