Update patch set 1
Patch Set 1: Code-Review+2

This basically allows the unit tests to be run in parallel.
It has been approved on VPP meeting, so +2.

But, what is the reason for not using the new way for every job? Are parallel tests somehow less reliable? Then we probably do not want them to affect the merge job.

Depending on the answer to that, the old script should be either deleted, or unified with the new one (e.g. have a single script but set an environment variable to control its behavior) as the difference is only one line (if I recall correctly).

Not pressing the submit button myself, as I do not know whether this will make things better or worse.

Patch-set: 1
Reviewer: Gerrit User 1263 <1263@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>
Label: Code-Review=+2