Update patch set 2
Patch Set 2: Code-Review-1

In other words it does not make sense at all.
I do not want csit-* jobs to be CRC checking. We need freedom in developing new features and debug _any_ version of VPP (This is huge problem of sub-related downstream projects I am working on, or on new tests that we are about to develop). I suggest to disable CRC by default in Constant.py and make only jobs that require it to enable it via Jenkins/VAR

Patch-set: 2