Update patch set 2
Patch Set 2: Code-Review-1

> I suggest to disable CRC by default in Constant.py and make only jobs that require it to enable it via Jenkins/VAR

Ok. It will reduce the amount of editing needed in ci-management. Will do in next change.
The text below explains why csit-perftest triggered jobs require the checks.

> We need freedom in developing new features and debug _any_ version of VPP

That is why we have a configurable value in Constants.py, so during development the contributor can chose to tolerate CRC mismatches (risking mysterious failures when testing with incompatible VPP version).
But committers still have to check that CSIT CRC list is updated (after the mysterious failures have been fixed) to the current stable_vpp before merging the Change.

> use warning instead of failing for crc check

I already forget to check pylint output most of the time. We may be careful enough when reviewing a Change that specifically supports a VPP API change (although [1] shows we are not), but will your warning be loud enough to prevent [2] from being merged if their edits to supported_crcs.yaml are not correct?

[1] https://gerrit.fd.io/r/c/csit/+/21922
[2] https://gerrit.fd.io/r/c/csit/+/20826

Patch-set: 2
Label: Code-Review=-1