Update patch set 3
Patch Set 3:

> how difficult it would be to add VPP CRC job logic into our tox?

Tox is independent of VPP build, so perhaps a standalone job would be better.
Download stable_vpp (or build a vpp), extract .api.json files, make sure one of CRC collections is compatible.

The main point of doing CRC checks during testing is to make sure every executed command (and received response) is in the CRC list.
If we had a good regular expression, we can perhaps identify all used commands, and add a tox checker (which also checks the corresponding responses are in the list).

The two jobs together would be enough, so we can start skipping the CRC checks during test execution.

Patch-set: 3