Update patch set 1
Patch Set 1:

This was discussed and agreed in the vpp meeting 10/15.  The only open action from that was for me to verify that csit was not using the images.   I verified and closed that AI that csit is not using any image pulled from nexus in the csit meeting 10/16

Patch-set: 1