CSIT: Add executor value to job description
For some jobs, it is not entirely clear which executor is used.
People verifying changes related to executors
previously needed to click job configuration to see it.
Now they only need to look at job description.
If a job does not list the executor in its description,
it means it does not use anything fancy.

Change-Id: Idbfdc015a9989b8ff65a8a0f0ca6593c16b2683f
Signed-off-by: Vratko Polak <vrpolak@cisco.com>