Update patch set 1
Patch Set 1:

> testing of something not merged is not 1:1 with periodic job.

It does not have to be 1:1, especially when the point is to show what the difference is.
Specifically, my use case is to show which tests are run after changing suite generatore (and periodic job tag expressions).

> you can use regular jobs for that

I guess you mean non-periodic ones. Results of periodic jobs are processed by trending, so we do not want to pollute them.

> Otherwise use regular job with refspec value to get the same.

Have you seen how long the tag expressions for periodic jobs are? I can create a long line, but then reviewers need to trust (or re-check) I did not make an error when pasting.

> unwanted autotrigger

The child of this Change is [1]. Not to be merged, but removes timer triggers, so creating periodic jobs on Sandbox should be safe.

Overall, we can keep this Change unmerged (as I only need this code locally to create such jobs on Sandbox), if you think merging this would increase the risk of reserving testbeds for wrong reasons.

[1] https://gerrit.fd.io/r/c/ci-management/+/24012

Patch-set: 1