CSIT: Make script inclusion more readable
Historically, scripts names started with include-raw-csit- prefix.
That was not only long and superfluous, but also confusing,
as the actual was if including uses
"include-raw-escape", which is a command behaving differently
tha "include-raw", suggested by the previous name.
Also, sometimes inclusion pointed to jjb/csit/,
which is not needed as it is the current directory.

This Change shortens script names to the part
that actually says something about the script,
and removes the directory part from the include lines.

Change-Id: Ifcfad5833577d78bf9105f494d5bf6187abc2a68
Signed-off-by: Vratko Polak <vrpolak@cisco.com>