VPP: Make script inclusion more readable
Historically, scripts names started with include-raw-vpp- prefix.
That was not only long and superfluous, but also confusing,
as the actual way of including uses
"include-raw-escape", which is a command behaving differently
tha "include-raw", suggested by the previous name.

Also, some include lines contain only file name,
relying on JJB finding the correct directory.

This Change shortens script names to the part
that actually says something about the script,
and prepends "./" to make it explicit where the script is.

+ Delete include-gcc8-prefer.sh as it is orphaned since
  https://gerrit.fd.io/r/c/ci-management/+/20443

Change-Id: I48920542ff20810cc87fb01844236350348bae59
Signed-off-by: Vratko Polak <vrpolak@cisco.com>