CSIT: Make script inclusion more readable
Historically, scripts names started with include-raw-csit- prefix.
That was not only long and superfluous, but also confusing,
as the actual way of including uses
"include-raw-escape", which is a command behaving differently
than "include-raw", suggested by the previous name.

This Change shortens script names to the part
that actually says something about the script.

+ Move scripts to jjb/scripts/csit/.
- "Absolute" paths are not used, as ci-man root is not in serarch path.
+ Put each path on a separate line to avoid long lines.
+ Update copyright years, add missing copyright notices.

Change-Id: Ifcfad5833577d78bf9105f494d5bf6187abc2a68
Signed-off-by: Vratko Polak <vrpolak@cisco.com>