VPP: Make script inclusion more readable
Historically, scripts names started with include-raw-vpp- prefix.
That was not only long and superfluous, but also confusing,
as the actual way of including uses
"include-raw-escape", which is a command behaving differently
than "include-raw", suggested by the previous name.

This Change shortens script names to the part
that actually says something about the script.

+ Delete include-gcc8-prefer.sh as it is orphaned since
  https://gerrit.fd.io/r/c/ci-management/+/20443
+ Move the scripts to jjb/scripts/vpp/.
- "absolute" paths are not used, as ci-man root is not in search path.
+ Put each path on a separate line to avoid long lines.
+ Put each include in a separate shell to make sure
  return code of each script is checked.
+ Update copyright year where present.
- Copyright notices are not added when not present.

Change-Id: I48920542ff20810cc87fb01844236350348bae59
Signed-off-by: Vratko Polak <vrpolak@cisco.com>