Break ci-management jobs group down
While trying to implement / revert some of the work for the Jenkins
managed configuration jobs we've run into some strange problems.
Unfortunately there is currently a bug that we're hitting in global-jjb
that is causing all configurations to disapear.

This change switches from using the job group to defining just the jobs
we want in the system right now so that we don't keep running into this
bug / problem. It also allows us to continue testing configuration
changes without fear of breaking the system.

Change-Id: I49ffd839762de88b51db0d9d70bbcff1624c395e
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>