Update patch set 4
Patch Set 4: Code-Review-1

> Did you test the jobs on the sandbox? In the past, we've had job failures as a result of using the different publisher.

Yes, this was created before the failures, so it worked for a brief time. But then something (I guess [0]) happened and the macro stopped working. I still use this change to monitor it is still not working.

Waiting for the macro to start working again.

[0] https://gerrit.fd.io/r/c/ci-management/+/23526

Patch-set: 4
Label: Code-Review=-1