Remove project specific packages
This change is remove the project specific packages that were originally
in baseline as a requirement primarily for RPM_DPDK, DEB_DPDK and VPP.

The RPM_DPDK and DEB_DPDK projects are inactive/archived and VPP
manages their own Docker images.

I'll add required packages back as they are found through testing
projects that are building on OpenStack images or as requested via the
projects.

Signed-off-by: Vanessa Rene Valderrama <vvalderrama@linuxfoundation.org>
Change-Id: I207b02854622f676288954251d7446eeb70e150e