Update patch set 3
Patch Set 3: Code-Review+2

So this is basically the same as csit-vpp-perf-verify, but with different trigger (and without CRC checking)? I guess the motivation is to have a job with history consisting of "official" results, and to keep the verify job for testing CSIT changes that are not official yet.
Ok, why not.

Patch-set: 3
Reviewer: Gerrit User 1263 <1263@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>
Label: Code-Review=+2