Update patch set 1
Patch Set 1:

> I meant script wise:
> include-raw-deploy-archives.sh
> include-raw-telemetry.sh

You mean execute two independent scripts (just in the opposite order, so telemetry output gets archived), right?
Why not, as long as the telemetry script never messes with the job run status (e.g. it never fails).

But even then, we will probably want to execute both scripts from the same macro, as there is no reason to archive without telemetry. (And the combined macro can call a telemetry-only macro, in case we want to gather telemetry in multiple points of a run.)

We can keep an archive only macro (for modularity purposes), but in that case we should rename fdio-infra-shiplogs to some new name, so old open Changes do not create jobs with archiving without telemetry. And of course replace occurrences of fdio-infra-shiplogs with the new combined macro.

Patch-set: 1