Update patch set 1
Patch Set 1:

> lf-infra-publish global-jjb macro

See my comments in [1] for why we do not use that in most jobs.
Vanessa has a task listed in [2] to make it work.
I guess it could be just a question of installing the correct dependencies into executor (docker) images we use.

Does global-jjb have a Jira or something similar, to publicly track any comments from Vanessa and you?

[1] https://gerrit.fd.io/r/c/ci-management/+/26735
[2] https://wiki.fd.io/view/Nomad_Operations_and_Planning

Patch-set: 1