Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Do we want to create a job for gcc instead?

I think that gcc is still (or going to revert to) being used for building arm64 packages, so I don't think that is strictly necessary.  I think the original motivation of having the clang build job was to catch more compile time errors, than ensuring portability (but I may be wrong).

Probably not a bad idea if none of the packages are built using gcc. But that also raises the question of which version of gcc -- 9, 10, ???

Patch-set: 1