Update patch set 3
Patch Set 3:

Ok the teams room really is mostly a temporary thing so that I (and whoever else wants) can real-time jump in with folks and discuss the challenges they are facing when making a new API, so that we don't just leave them in the void.

After a couple of weeks I aim to remove the notification about the successful run, and then in a 6-9 months or a year or so maybe even for the failed checks.

Patch-set: 3