Update patch set 17
Patch Set 17:

(1 comment)

> Patch Set 17: Code-Review-1
> 
> Please don't merge this patch until we have input from Release Management. We should be working on a solution to follow the release workflow. Pushing directly to the release repository is not following best practice.

We have had a discussion on this issue on the list and on slack.
This is a solution that was fine for Andrew. Please double check there.

The alternative is to create as many master repos as the number of projects.
All options have been tabled on the list though and we did not hear from you.
Nonetheless, I believe you are right but we need actions from LF to execute.
Please reply on the list with a solution. We'll follow up there.

Patch-set: 17
Reviewer: Gerrit User 789 <789@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>