Update patch set 20
Patch Set 20:

> https://jenkins.fd.io/sandbox/job/release-repo-cleanup-ubuntu1804/16/console

Ok. So when I look at libmemif, I see the highest version to promote is libmemif_19.08-202-release_amd64.deb
But when looking at [5] (I did not find a way to hide ARM packages from the search output), I see 6 (not just 5) versions higher than 202: 203, 205, 206, 207, 209, 210.

Also, the smallest version of libmemif to promote is libmemif_19.08-17-release_amd64.deb
but packagecloud also contains libmemif_19.08_amd64.deb
It probably does not match the regexp, so it is ok for the script to keep it.

[5] https://packagecloud.io/app/fdio/release/search?dist=ubuntu%2Fbionic&filter=all&page=5&q=libmemif_19.08

Patch-set: 20