Update patch set 21
Patch Set 21:

> Patch Set 20:
> 
> > https://jenkins.fd.io/sandbox/job/release-repo-cleanup-ubuntu1804/16/console
> 
> Ok. So when I look at libmemif, I see the highest version to promote is libmemif_19.08-202-release_amd64.deb
> But when looking at [5] (I did not find a way to hide ARM packages from the search output), I see 6 (not just 5) versions higher than 202: 203, 205, 206, 207, 209, 210.
> 
> Also, the smallest version of libmemif to promote is libmemif_19.08-17-release_amd64.deb
> but packagecloud also contains libmemif_19.08_amd64.deb
> It probably does not match the regexp, so it is ok for the script to keep it.
> 
> [5] https://packagecloud.io/app/fdio/release/search?dist=ubuntu%2Fbionic&filter=all&page=5&q=libmemif_19.08

Yes the problem of libmemif is that the versioning is not uniform:

libmemif:
16:39:34   Installed: (none)
16:39:34   Candidate: 20.05
16:39:34   Version table:
16:39:34      20.05 500
16:39:34         500 https://packagecloud.io/fdio/release/ubuntu bionic/main amd64 Packages
16:39:34      20.01 500
16:39:34         500 https://packagecloud.io/fdio/release/ubuntu bionic/main amd64 Packages
16:39:34      19.08-210-release 500
16:39:34         500 https://packagecloud.io/fdio/release/ubuntu bionic/main amd64 Packages

So the regex starts considering from 19.08-210-release, keeping effectively one additional package, since 19.08-210-release is not repeated. However for libmemif my opinion is that we should keep only the 20.05, 20.01 and 19.08 packages, since it is not really part of the hicn package suite and there is not need to keep snapshots, but only the versions matching the corresponding VPP release. I will remove it from the package list and I will open a case to delete the additional packages with version 19.08-X.

Patch-set: 21
Reviewer: Gerrit User 794 <794@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>