Update patch set 1
Patch Set 1:

That error with VPP not starting (dpdk_plugin.so: undefined symbol: va_unformat) might be because of out-of-date vpp-ext-deps. We're not downloading and updating it in CSIT scripts:
Didn't find '21.01-rc0~230-g6acb7a03c~b134' among 'vpp-ext-deps' versions.

How should we handle this? Do we want to update all docker images (I guess only all, but only the corresponding branches) everytime a new vpp-ext-deps is published or should we fix this in CSIT scripts?

Patch-set: 1