Update patch set 1
Patch Set 1:

One thing that super slow for some reason is this:
14:49:16 Robot results publisher started...
14:49:16 -Parsing output xml:
14:49:18 Done!
14:49:18 -Copying log files to build dir:
14:59:30 Done!
14:59:30 -Assigning results to build:
14:59:30 Done!
14:59:30 -Checking thresholds:
14:59:30 Done!
14:59:30 Done publishing Robot results.

Any idea why this copying is so slow? Could it be because it's running in Sandbox? And older run of the old job is much faster: https://logs.fd.io/production/vex-yul-rot-jenkins-1/csit-vpp-device-master-ubuntu1804-1n-tx2-csit-verify-hourly/9144/console-timestamp.log.gz
12:41:26  Robot results publisher started...
12:41:26  -Parsing output xml:
12:42:28  Done!
12:42:28  -Copying log files to build dir:
12:43:39  Done!
12:43:39  -Assigning results to build:
12:43:39  Done!
12:43:39  -Checking thresholds:
12:43:39  Done!
12:43:39  Done publishing Robot results.

Patch-set: 1